chore: remove interfaces from project

+++ Models/Sinistro.cs
@@ -8,7 +8,7 @@ public class Sinistro
     public Guid OficinaId { get; set; }
     public Guid ReguladorId { get; set; }
     public string Chassi { get; set; } = "";
-    public double ValorReparo { get; set; }
+    public decimal ValorReparo { get; set; }
     public bool Salvado { get; set; }
     public SinistroStatus Status { get; set; }
     public DateTime DataSinistro { get; set; }
